Split into tables sub-module
